+++ libcloud/compute/ssh.py
@@ -14,8 +14,9 @@
 # limitations under the License.
 
 """
-Wraps multiple ways to communicate over SSH
+Wraps multiple ways to communicate over SSH.
 """
+
 have_paramiko = False
 
 try:
@@ -32,6 +33,7 @@ import os
 import time
 import subprocess
 import logging
+import warnings
 
 from os.path import split as psplit
 from os.path import join as pjoin
@@ -50,7 +52,7 @@ class BaseSSHClient(object):
     """
 
     def __init__(self, hostname, port=22, username='root', password=None,
-                 key=None, timeout=None):
+                 key=None, key_files=None, timeout=None):
         """
         :type hostname: ``str``
         :keyword hostname: Hostname or IP address to connect to.
@@ -66,14 +68,24 @@ class BaseSSHClient(object):
                            to unlock a private key if a password protected key
                            is used.
 
-        :type key: ``str`` or ``list``
-        :keyword key: A list of paths to the private key files to use.
+        :param key: Deprecated in favor of ``key_files`` argument.
+
+        :type key_files: ``str`` or ``list``
+        :keyword key_files: A list of paths to the private key files to use.
         """
+        if key is not None:
+            message = ('You are using deprecated "key" argument which has '
+                       'been replaced with "key_files" argument')
+            warnings.warn(message, DeprecationWarning)
+
+            # key_files has precedent
+            key_files = key if not key_files else key_files
+
         self.hostname = hostname
         self.port = port
         self.username = username
         self.password = password
-        self.key = key
+        self.key_files = key_files
         self.timeout = timeout
 
     def connect(self):
@@ -116,8 +128,8 @@ class BaseSSHClient(object):
         :type path: ``str``
         :keyword path: File path on the remote node.
 
-        :return: True if the file has been successfully deleted, False
-                 otherwise.
+        :return: True if the file has been successfully deleted,
+                 False otherwise.
         :rtype: ``bool``
         """
         raise NotImplementedError(
@@ -139,8 +151,8 @@ class BaseSSHClient(object):
         """
         Shutdown connection to the remote node.
 
-        :return: True if the connection has been successfully closed, False
-                 otherwise.
+        :return: True if the connection has been successfully closed,
+                 False otherwise.
         :rtype: ``bool``
         """
         raise NotImplementedError(
@@ -165,7 +177,7 @@ class ParamikoSSHClient(BaseSSHClient):
     A SSH Client powered by Paramiko.
     """
     def __init__(self, hostname, port=22, username='root', password=None,
-                 key=None, timeout=None):
+                 key=None, key_files=None, key_material=None, timeout=None):
         """
         Authentication is always attempted in the following order:
 
@@ -177,8 +189,19 @@ class ParamikoSSHClient(BaseSSHClient):
         - Plain username/password auth, if a password was given (if password is
           provided)
         """
-        super(ParamikoSSHClient, self).__init__(hostname, port, username,
-                                                password, key, timeout)
+        if key_files and key_material:
+            raise ValueError(('key_files and key_material arguments are '
+                              'mutually exclusive'))
+
+        super(ParamikoSSHClient, self).__init__(hostname=hostname, port=port,
+                                                username=username,
+                                                password=password,
+                                                key=key,
+                                                key_files=key_files,
+                                                timeout=timeout)
+
+        self.key_material = key_material
+
         self.client = paramiko.SSHClient()
         self.client.set_missing_host_key_policy(paramiko.AutoAddPolicy())
         self.logger = self._get_and_setup_logger()
@@ -193,10 +216,13 @@ class ParamikoSSHClient(BaseSSHClient):
         if self.password:
             conninfo['password'] = self.password
 
-        if self.key:
-            conninfo['key_filename'] = self.key
+        if self.key_files:
+            conninfo['key_filename'] = self.key_files
 
-        if not self.password and not self.key:
+        if self.key_material:
+            conninfo['pkey'] = self._get_pkey_object(key=self.key_material)
+
+        if not self.password and not (self.key_files or self.key_material):
             conninfo['allow_agent'] = True
             conninfo['look_for_keys'] = True
 
@@ -345,6 +371,23 @@ class ParamikoSSHClient(BaseSSHClient):
         self.client.close()
         return True
 
+    def _get_pkey_object(self, key):
+        """
+        Try to detect private key type and return paramiko.PKey object.
+        """
+
+        for cls in [paramiko.RSAKey, paramiko.DSSKey, paramiko.ECDSAKey]:
+            try:
+                key = cls.from_private_key(StringIO(key))
+            except paramiko.ssh_exception.SSHException:
+                # Invalid key, try other key type
+                pass
+            else:
+                return key
+
+        msg = 'Invalid or unsupported key type'
+        raise paramiko.ssh_exception.SSHException(msg)
+
 
 class ShellOutSSHClient(BaseSSHClient):
     """
@@ -355,9 +398,13 @@ class ShellOutSSHClient(BaseSSHClient):
     """
 
     def __init__(self, hostname, port=22, username='root', password=None,
-                 key=None, timeout=None):
-        super(ShellOutSSHClient, self).__init__(hostname, port, username,
-                                                password, key, timeout)
+                 key=None, key_files=None, timeout=None):
+        super(ShellOutSSHClient, self).__init__(hostname=hostname,
+                                                port=port, username=username,
+                                                password=password,
+                                                key=key,
+                                                key_files=key_files,
+                                                timeout=timeout)
         if self.password:
             raise ValueError('ShellOutSSHClient only supports key auth')
 
@@ -403,8 +450,8 @@ class ShellOutSSHClient(BaseSSHClient):
     def _get_base_ssh_command(self):
         cmd = ['ssh']
 
-        if self.key:
-            cmd += ['-i', self.key]
+        if self.key_files:
+            cmd += ['-i', self.key_files]
 
         if self.timeout:
             cmd += ['-oConnectTimeout=%s' % (self.timeout)]